requirements.txt
tweepy
spaCy
folium
geopy
tqdm
geograpy3
python-dotenv>=0.19.2
schedule
beautifulsoup4
pytest
